// frontend/src/api/generated/models/notificationDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */
import type { NotificationType } from './notificationType';

export interface NotificationDto {
  id?: string;
  recipientUserId?: string;
  type?: NotificationType;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  createdAtUtc?: string;
  /** @nullable */
  readAtUtc?: string | null;
}