// frontend/src/api/generated/models/notificationType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  InviteReceived: 'InviteReceived',
  InviteAccepted: 'InviteAccepted',
  PlanUpdated: 'PlanUpdated',
  CommentAdded: 'CommentAdded',
  ReportReady: 'ReportReady',
  PermissionChanged: 'PermissionChanged',
  HealthFlagRaised: 'HealthFlagRaised',
  AccessRevoked: 'AccessRevoked',
} as const;